Revert "Refactor SystemController into separate components"
