--- cmd-file-service/src/main/java/de/zeroco/core/config/SwaggerConfig.java
@@ -1,24 +1,0 @@
-package de.zeroco.core.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@Configuration
-@EnableSwagger2
-public class SwaggerConfig {
-
-    @Bean
-    public Docket api() {
-        return new Docket(DocumentationType.OAS_30)
-            .select()
-            // Scan the broader 'de.zeroco.core' package to include all controllers
-            .apis(RequestHandlerSelectors.basePackage("de.zeroco.core"))
-            .paths(PathSelectors.any())
-            .build();
-    }
-}
