+++ cmd-file-service/src/main/java/de/zeroco/core/security/SecuredOperation.java
@@ -1,0 +1,15 @@
+package de.zeroco.core.security;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Marker annotation for operations that may require security checks (e.g., @PreAuthorize).
+ * This allows security configurations to target these specific methods.
+ */
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface SecuredOperation {
+}
