--- cmd-file-service/src/main/resources/application.properties
@@ -1,12 +1,0 @@
-# Security Configuration
-api.security.username=user
-api.security.password=password
-
-# Application Feature Paths
-app.execute.sandbox-path=sandbox
-app.download.base-path=downloads
-app.logs.base-path=logs_data
-app.logs.default-file=application.log
-
-# Default Spring Boot server port (can be overridden by user)
-server.port=8080
